refactor: 에러 상태 관리를 useErrorContext를 사용하도록 수정
